new alloc counter tests
