+++ IntegrationTests/allocation-counter-tests-framework/template/AtomicCounter/Package.swift
