+++ IntegrationTests/allocation-counter-tests-framework/template/HookedFunctionsDoHook/Package.swift
