+++ IntegrationTests/allocation-counter-tests-framework/template/Sources/bootstrapDoHook/main.c
