+++ IntegrationTests/tests_04_performance/test_01_resources/test_1000_reqs_1_conn.swift
@@ -1,0 +1,21 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the SwiftNIO open source project
+//
+// Copyright (c) 2017-2019 Apple Inc. and the SwiftNIO project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of SwiftNIO project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+func run(identifier: String) {
+    measure(identifier: identifier) {
+        let numberDone = try! doRequests(group: group, number: 1000)
+        precondition(numberDone == 1000)
+        return numberDone
+    }
+}
